prepare-inference/requirements.txt
# ── Core ML ────────────────────────────────────────────────────────────────
torch>=2.1.0
transformers>=4.40.0
peft==0.18.1
accelerate>=0.27.0
bitsandbytes>=0.43.0
safetensors>=0.4.0
sentencepiece
protobuf

# ── RAG ────────────────────────────────────────────────────────────────────
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0          # use faiss-gpu if you have a GPU index that fits VRAM
pandas>=2.0.0

# ── Khmer NLP ───────────────────────────────────────────────────────────────
khmer-nltk>=0.1.5

# ── API ────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
httpx

# ── Frontend ───────────────────────────────────────────────────────────────
streamlit>=1.35.0
requests>=2.31.0

# ── Utilities ──────────────────────────────────────────────────────────────
numpy>=1.26.0
python-dotenv
